[jacodb-ifds] Minor refactoring to improve logic and Java API
