+++ jacodb-analysis/src/main/kotlin/org/jacodb/analysis/AnalysisMain.kt
@@ -17,21 +17,14 @@
 package org.jacodb.analysis
 import kotlinx.serialization.Serializable
 import mu.KLogging
-import org.jacodb.analysis.analyzers.AliasAnalyzerFactory
-import org.jacodb.analysis.analyzers.NpeAnalyzerFactory
-import org.jacodb.analysis.analyzers.NpePrecalcBackwardAnalyzerFactory
-import org.jacodb.analysis.analyzers.SqlInjectionAnalyzerFactory
-import org.jacodb.analysis.analyzers.SqlInjectionBackwardAnalyzerFactory
-import org.jacodb.analysis.analyzers.TaintAnalysisNode
-import org.jacodb.analysis.analyzers.TaintAnalyzerFactory
-import org.jacodb.analysis.analyzers.TaintBackwardAnalyzerFactory
-import org.jacodb.analysis.analyzers.TaintNode
-import org.jacodb.analysis.analyzers.UnusedVariableAnalyzerFactory
-import org.jacodb.analysis.engine.IfdsBaseUnitRunner
-import org.jacodb.analysis.engine.SequentialBidiIfdsUnitRunner
+import org.jacodb.analysis.engine.IfdsUnitManager
+import org.jacodb.analysis.engine.IfdsUnitRunner
+import org.jacodb.analysis.engine.Summary
 import org.jacodb.analysis.engine.TraceGraph
+import org.jacodb.analysis.engine.UnitResolver
+import org.jacodb.analysis.graph.newApplicationGraphForAnalysis
 import org.jacodb.api.JcMethod
-import org.jacodb.api.cfg.JcExpr
+import org.jacodb.api.analysis.JcApplicationGraph
 import org.jacodb.api.cfg.JcInst
 
 /**
@@ -83,44 +76,41 @@ typealias AnalysesOptions = Map<String, String>
 @Serializable
 data class AnalysisConfig(val analyses: Map<String, AnalysesOptions>)
 
-val UnusedVariableRunner = IfdsBaseUnitRunner(UnusedVariableAnalyzerFactory)
-
-fun newSqlInjectionRunner(maxPathLength: Int = 5) = SequentialBidiIfdsUnitRunner(
-    IfdsBaseUnitRunner(SqlInjectionAnalyzerFactory(maxPathLength)),
-    IfdsBaseUnitRunner(SqlInjectionBackwardAnalyzerFactory(maxPathLength)),
-)
-
-fun newNpeRunner(maxPathLength: Int = 5) = SequentialBidiIfdsUnitRunner(
-    IfdsBaseUnitRunner(NpeAnalyzerFactory(maxPathLength)),
-    IfdsBaseUnitRunner(NpePrecalcBackwardAnalyzerFactory(maxPathLength)),
-)
-
-fun newAliasRunner(
-    generates: (JcInst) -> List<TaintAnalysisNode>,
-    sanitizes: (JcExpr, TaintNode) -> Boolean,
-    sinks: (JcInst) -> List<TaintAnalysisNode>,
-    maxPathLength: Int = 5
-) = IfdsBaseUnitRunner(AliasAnalyzerFactory(generates, sanitizes, sinks, maxPathLength))
-
-fun newTaintRunner(
-    isSourceMethod: (JcMethod) -> Boolean,
-    isSanitizeMethod: (JcMethod) -> Boolean,
-    isSinkMethod: (JcMethod) -> Boolean,
-    maxPathLength: Int = 5
-) = SequentialBidiIfdsUnitRunner(
-    IfdsBaseUnitRunner(TaintAnalyzerFactory(isSourceMethod, isSanitizeMethod, isSinkMethod, maxPathLength)),
-    IfdsBaseUnitRunner(TaintBackwardAnalyzerFactory(isSourceMethod, isSinkMethod, maxPathLength))
-)
-
-fun newTaintRunner(
-    sourceMethodMatchers: List<String>,
-    sanitizeMethodMatchers: List<String>,
-    sinkMethodMatchers: List<String>,
-    maxPathLength: Int = 5
-) = SequentialBidiIfdsUnitRunner(
-    IfdsBaseUnitRunner(TaintAnalyzerFactory(sourceMethodMatchers, sanitizeMethodMatchers, sinkMethodMatchers, maxPathLength)),
-    IfdsBaseUnitRunner(TaintBackwardAnalyzerFactory(sourceMethodMatchers, sinkMethodMatchers, maxPathLength))
-)
-
-
 internal val logger = object : KLogging() {}.logger
+
+/**
+ * This is the entry point for every analysis.
+ * Calling this function will find all vulnerabilities reachable from [methods].
+ *
+ * @param graph instance of [JcApplicationGraph] that provides mixture of CFG and call graph
+ * (called supergraph in RHS95).
+ * Usually built by [newApplicationGraphForAnalysis].
+ *
+ * @param unitResolver instance of [UnitResolver] which splits all methods into groups of methods, called units.
+ * Units are analyzed concurrently, one unit will be analyzed with one call to [IfdsUnitRunner.run] method.
+ * In general, larger units mean more precise, but also more resource-consuming analysis, so [unitResolver] allows
+ * to reach compromise.
+ * It is guaranteed that [Summary] passed to all units is the same, so they can share information through it.
+ * However, the order of launching and terminating analysis for units is an implementation detail and may vary even for
+ * consecutive calls of this method with same arguments.
+ *
+ * @param ifdsUnitRunner an [IfdsUnitRunner] instance that will be launched for each unit.
+ * This is the main argument that defines the analysis.
+ *
+ * @param methods the list of method for analysis.
+ * Each vulnerability will only be reported if it is reachable from one of these.
+ *
+ * @param timeoutMillis the maximum time for analysis.
+ * Note that this does not include time for precalculations
+ * (like searching for reachable methods and splitting them into units) and postcalculations (like restoring traces), so
+ * the actual running time of this method may be longer.
+ */
+fun runAnalysis(
+    graph: JcApplicationGraph,
+    unitResolver: UnitResolver<*>,
+    ifdsUnitRunner: IfdsUnitRunner,
+    methods: List<JcMethod>,
+    timeoutMillis: Long = Long.MAX_VALUE
+): List<VulnerabilityInstance> {
+    return IfdsUnitManager(graph, unitResolver, ifdsUnitRunner, methods, timeoutMillis).analyze()
+}
